Rework protocol validators
