+++ Tzkt.Sync/Protocols/Handlers/Proto7/Validation/Validator.cs
@@ -1,6 +1,6 @@
 ﻿namespace Tzkt.Sync.Protocols.Proto7
 {
-    class Validator : Proto6.Validator
+    class Validator : Proto1.Validator
     {
         public Validator(ProtocolHandler protocol) : base(protocol) { }
     }
